requirements.txt
fire
opensmile
pandas
pyannote.core
pydub
